requirements.txt
pdfplumber==0.10.4
python-docx==1.1.2
pyresparser==1.0.6  # AI parsing with NLP
sentence-transformers==3.0.1  # Semantic scoring with embeddings